added static casts into uint32_t for rgb values in SetPixel()

+++ src/main_utils.cpp
@@ -48,7 +48,7 @@ int RandInt(int rangeMin, int rangeMax) {
 
 void SetPixel(uint32_t* buffer, int width, int height, int x, int y, uint8_t r, uint8_t g, uint8_t b, uint8_t a) {
     if (x >= 0 && x < width && y >= 0 && y < height) {
-        uint32_t color = (r << 24) | (g << 16) | (b << 8) | a; // translates base-255 into proper color format
+        uint32_t color = (static_cast<uint32_t>(r) << 24) | (static_cast<uint32_t>(g) << 16) | (static_cast<uint32_t>(b) << 8) | a; // translates base-255 into proper color format
         buffer[(y * width) + x] = color;
     }
 }
